Refactor test utility imports and update sonar exclusions for test utility files
