+++ .github/workflows/build.yml
@@ -2,7 +2,7 @@ name: CI
 on:
   push:
     branches:
-      - "main"
+      - 'main'
   pull_request:
 
 permissions: {}
