+++ demo/tsconfig.json
@@ -1,6 +1,7 @@
 {
     "extends": "../tsconfig.base.json",
     "compilerOptions": {
+        "rootDir": "..",
         "outDir": "./dist",
         "jsx": "react-jsx",
         "paths": {
