+++ sonar-project.properties
@@ -3,5 +3,5 @@ sonar.organization=powsybl-ci-github
 sonar.javascript.lcov.reportPaths=./coverage/lcov.info
 sonar.sources=src/,packages/network-map-layers/src/,packages/network-viewer-core/src/
 sonar.tests=src/,packages/network-map-layers/src/,packages/network-viewer-core/src/
-sonar.exclusions=**/*.test.ts,**/node_modules/**,**/dist/**,packages/network-map-layers/src/testUtils/**
+sonar.exclusions=**/*.test.ts,**/node_modules/**,**/dist/**,**/*.test.utils.ts
 sonar.test.inclusions=**/*.test.ts
